Use dev.cdevents as root for event types
According to the CloudEvents spec, https://github.com/cloudevents/spec/blob/v1.0.1/spec.md
the event type SHOULD be prefixed with a reverse-DNS name.
The prefix domain dictates the organisation which defines
the semantics of the event type.

Since CDEvents owns the `cdevents.dev` domain, and has a website
hosted on that domain, we should use that domain in the event
types as well.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>