Add change to the artifact packaged event
The artifact packaged event is typically produced by the build process,
and should provide details about the source the artifact was built from.

While working on the artifact change field, I realised that for the
change and branch events the id + source combination should also contain
the repository details, otherwise it wouldn't uniquely identify a
change.

I added a note in the change/branch event section and updated the
examples accordingly. I think there is still value in have a repository
field in those events, but I marked it as optional (it was mandatory
before).

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>