Add artifactId to service events
Add the artifactId to some of the service events:
- published
- upgraded
- rolledback

The model is that one service is associated to one artifact.
As already documented in the primer.md, this is a simplified model
and it may change in future versions.

The format of the artifactId in the examples is consistent with the
rest of the spec for now. All artifactIds will align to the purl
format in an upcoming pull request.

Fixes: #58

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>