Update service environment to be a ref
The service environment is a ref to an environment in the schemas
and it should be a ref, i.e. a combination of id and source because
the id is only unique when combined to the source.

As in other cases, if the source of the environment is the same as
the source of the main subject (the service), it may be omitted.

The spec only includes today the environmentId as string - update
it to be an object (reference to environment).

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>