Repo optional in branch and change schemas
The repository is an optional field in branch and change events,
both in the specification as well as in the SDK.

This is because the source includes the repository information already.
While there is value in specifying the repository reference separated
from the SCM system URL, it should not be mandatory to do so.

The schema files for branch and change events specify the repository
as a mandatory field. This was an oversight, which is fixed here.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>