Remove the version enums from schemas
Including the spec version enums in schemas is problematic, because
every time the spec is updated it implies that all the event schemas
must be updated as well which in turn requires the event type versions
to be updated too, which is confusing since not all events are updated
when a new spec version is made.

To solve that issue, we remove the enum from the version property, a let
it be a non-empty string. The SDKs can use the event type and version to
download the schema and validate events.

When a new spec release is made, the $id in the schema of all events
still have to be updated, but that should not cause the version on the
event type to be updated.

Fixes #87

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>