Add change.created description field
This commit adds a new optional field to the dev.cdevents.change.created
event.  This field, description, should be used to describe the change,
e.g. a GitHub PR.

An example of the description field could be something like

"This PR addresses a new bug that was introduced in some PR"

Signed-off-by: xibz <impactbchang@gmail.com>