Update testsuiterunqueued.json to replace url with uri
The testsuiterunqueued schema uses the field url under testsuite. Within the other testsuite run schemas, uri is used. This is also used in the markdown documents.

This change updates the testsuiterunqueued schema to bring it in line with the other testsuite's

This is in relation to: https://github.com/cdevents/spec/issues/160

Signed-off-by: Owen Bower Adams <owen+github@obowersa.net>