Refine the spelling configurartion
Split spelling check of json and markdown files.

For markdown add a context filter to ignore code blocks (inline and multiline).
Enable url and camel case modes.

Change the action to a different one, based on the same technology
(pyspelling + aspell) as the current one does not support the came case
mode.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>